+++ istioctl/pkg/install/pre-check.go
@@ -20,8 +20,6 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"regexp"
-	"strconv"
 	"strings"
 
 	"github.com/ghodss/yaml"
@@ -37,16 +35,11 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	"istio.io/istio/istioctl/pkg/clioptions"
+	"istio.io/istio/istioctl/pkg/install/k8sversion"
 	operator_istio "istio.io/istio/operator/pkg/apis/istio"
 	operator_v1alpha1 "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 	"istio.io/istio/operator/pkg/util"
 	"istio.io/istio/pilot/pkg/serviceregistry/kube/controller"
-)
-
-const (
-	// Minimum K8 version required to run latest version of Istio
-	// https://istio.io/docs/setup/platform-setup/
-	minK8SVersion = "1.16"
 )
 
 var (
@@ -99,12 +92,12 @@ func installPreCheck(istioNamespaceFlag string, restClientGetter genericclioptio
 	fmt.Fprintf(writer, "\n")
 	fmt.Fprintf(writer, "#2. Kubernetes-version\n")
 	fmt.Fprintf(writer, "-----------------------\n")
-	res, err := checkKubernetesVersion(v)
+	res, err := k8sversion.CheckKubernetesVersion(v)
 	if err != nil {
 		errs = multierror.Append(errs, err)
 		fmt.Fprint(writer, err)
 	} else if !res {
-		msg := fmt.Sprintf("The Kubernetes API version: %v is lower than the minimum version: "+minK8SVersion, v)
+		msg := fmt.Sprintf("The Kubernetes API version: %v is lower than the minimum version: "+k8sversion.MinK8SVersion, v)
 		errs = multierror.Append(errs, errors.New(msg))
 		fmt.Fprintf(writer, msg+"\n")
 	} else {
@@ -218,42 +211,6 @@ func installPreCheck(istioNamespaceFlag string, restClientGetter genericclioptio
 	fmt.Fprintf(writer, "\n")
 	return errs
 
-}
-
-func checkKubernetesVersion(versionInfo *version.Info) (bool, error) {
-	v, err := extractKubernetesVersion(versionInfo)
-	if err != nil {
-		return false, err
-	}
-	return parseVersion(minK8SVersion, 4) <= parseVersion(v, 4), nil
-}
-func extractKubernetesVersion(versionInfo *version.Info) (string, error) {
-	versionMatchRE := regexp.MustCompile(`^\s*v?([0-9]+(?:\.[0-9]+)*)(.*)*$`)
-	parts := versionMatchRE.FindStringSubmatch(versionInfo.GitVersion)
-	if parts == nil {
-		return "", fmt.Errorf("could not parse %q as version", versionInfo.GitVersion)
-	}
-	numbers := parts[1]
-	components := strings.Split(numbers, ".")
-	if len(components) <= 1 {
-		return "", fmt.Errorf("the version %q is invalid", versionInfo.GitVersion)
-	}
-	v := strings.Join([]string{components[0], components[1]}, ".")
-	return v, nil
-}
-func parseVersion(s string, width int) int64 {
-	strList := strings.Split(s, ".")
-	format := fmt.Sprintf("%%s%%0%ds", width)
-	v := ""
-	for _, value := range strList {
-		v = fmt.Sprintf(format, v, value)
-	}
-	var result int64
-	var err error
-	if result, err = strconv.ParseInt(v, 10, 64); err != nil {
-		return 0
-	}
-	return result
 }
 
 func checkCanCreateResources(c preCheckExecClient, namespace, group, version, name string) error {
